requirements.txt
tensorflow
tensorflow-datasets
imageio
pillow
numpy
